+++ packages/twenty-server/src/engine/api/rest/controllers/core.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Delete, Get, Post, Put, Req, Res } from '@nestjs/common';
 
 import { Request, Response } from 'express';
 
-import { CoreService } from 'src/engine/api/rest/core.service';
+import { CoreService } from 'src/engine/api/rest/services/core.service';
 import { cleanGraphQLResponse } from 'src/engine/api/rest/controllers/utils/api-rest.controller.utils';
 
 @Controller('rest/*')
