+++ packages/twenty-server/src/engine/api/rest/controllers/metadata.controller.ts
@@ -2,7 +2,7 @@ import { Controller, Get, Delete, Post, Put, Req, Res } from '@nestjs/common';
 
 import { Request, Response } from 'express';
 
-import { ApiRestMetadataService } from 'src/engine/api/rest/metadata.service';
+import { ApiRestMetadataService } from 'src/engine/api/rest/services/metadata.service';
 import { cleanGraphQLResponse } from 'src/engine/api/rest/controllers/utils/api-rest.controller.utils';
 
 @Controller('rest/metadata/*')
