+++ packages/twenty-server/src/engine/api/rest/services/__tests__/core.service.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { HttpService } from '@nestjs/axios';
 
-import { CoreService } from 'src/engine/api/rest/core.service';
+import { CoreService } from 'src/engine/api/rest/services/core.service';
 import { EnvironmentService } from 'src/engine/integrations/environment/environment.service';
 import { TokenService } from 'src/engine/core-modules/auth/services/token.service';
 import { CoreQueryBuilderFactory } from 'src/engine/api/rest/core-query-builder/core-query-builder.factory';
